+++ stdnet/lib/__init__.py
@@ -1,12 +1,12 @@
 #First try local
 try:
     from .hr import *
-    hasextensions = True
+    hasextensions = True    # pragma nocover
 except ImportError:
     # Try Global
     try:
         from hr import *
-        hasextensions = True
+        hasextensions = True    # pragma nocover
     except ImportError:
         hasextensions = False
         from .fallback import *