+++ stdnet/orm/base.py
@@ -302,13 +302,6 @@ of fields names and a list of field attribute names.'''
                         names.append(name)
                         atts.append(name)
         return names,atts
-
-    def multifields_ids_todelete(self, instance):
-        '''Return the list of ids of :class:`MultiField` belonging to *instance*
-which needs to be deleted when *instance* is deleted.'''
-        gen = (field.id(instance) for field in self.multifields\
-                                         if field.todelete())
-        return [fid for fid in gen if fid]
         
 
 class autoincrement(object):