+++ stdnet/orm/mapper.py
@@ -100,7 +100,7 @@ on the same redis instance).'''
     if model in _GLOBAL_REGISTRY:
         if not ignore_duplicates:  
             raise AlreadyRegistered(
-                        'Model {0} is already registered'.format(meta))
+                    'Model {0} is already registered'.format(model._meta))
         else:
             return
     backend = getdb(backend_uri = backend, **params)
@@ -228,14 +228,5 @@ It return s a list of registered models.'''
     return list(register_application_models(applications,**kwargs))
 
 
-def register_query_for_model(model, query_function):
-    _SPECIAL_QUERIES[model] = query_function
-    
-def get_model_query(model):
-    return _SPECIAL_QUERIES.get(model)
-        
-
 _GLOBAL_REGISTRY = set()
-
-_SPECIAL_QUERIES = {}