+++ tests/regression/cache.py
@@ -1,0 +1,7 @@
+from stdnet import test, getcache
+
+class TestCache(test.TestCase):
+    
+    def testSimple(self):
+        c = getcache()
+        self.assertTrue(c)