+++ tests/regression/me.py
@@ -1,9 +1,10 @@
-from stdnet.test import TestCase
-
+from stdnet import test
+from stdnet.conf import settings
+from stdnet.lib.redis import ConnectionError
 import stdnet as me
 
 
-class TestInitFile(TestCase):
+class TestInitFile(test.TestCase):
 
     def test_version(self):
         self.assertTrue(me.VERSION)
@@ -14,3 +15,11 @@ class TestInitFile(TestCase):
     def test_meta(self):
         for m in ("__author__", "__contact__", "__homepage__", "__doc__"):
             self.assertTrue(getattr(me, m, None))
+            
+    def testSettings(self):
+        db = settings.DEFAULT_BACKEND
+        try:
+            settings.DEFAULT_BACKEND = 'redis://dksnkdcnskcnskcn:6379?db=7'
+            self.assertRaises(ConnectionError, settings.redis_status)
+        finally:
+            settings.DEFAULT_BACKEND = db