+++ tests/regression/query/meta.py
@@ -18,7 +18,15 @@ class TestMeta(FinanceTest):
         self.assertEqual(list(empty),[])
         self.assertEqual(empty.executed,True)
         self.assertEqual(empty.construct(),empty)
-        
+        self.assertEqual(empty.items(),[])
+        self.assertEqual(empty.all(),[])
+        self.assertEqual(empty.intersect(Instrument.objects.query()),empty)
+        all = Instrument.objects.query()
+        all2 = empty.union(all)
+        self.assertEqual(set(all),set(all2))
+        q = Instrument.objects.filter(ccy__in = ())
+        self.assertEqual(q.count(),0)
+                
     def testProperties(self):
         query = Instrument.objects.query()
         self.assertFalse(query.executed)
@@ -29,5 +37,10 @@ class TestMeta(FinanceTest):
         query = query.get_field('id')
         query2 = query.get_field('id')
         self.assertEqual(query,query2)
+        
+    def testFilterError(self):
+        query = Instrument.objects.filter(whoaaaaa = 'foo')
+        self.assertRaises(QuerySetError, query.all)
+